requirements.txt
Django >= 3.0.7,<3.1a1
djangorestframework==3.11.0
psycopg2 == 2.8.5
Pillow>=5.3.0,<5.4.0
django-ckeditor == 6.0.0
django-cleanup==5.1.0
uwsgi>=2.0.18,<2.1.0

